+++ SeaBot/Data/Definitions/DealerDefenitions.cs
@@ -12,7 +12,7 @@
 // GNU General Public License for more details.
 // 
 // You should have received a copy of the GNU General Public License
-// along with this program.  If not, see <http://www.gnu.org/licenses/>.
+// aint with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 using System.Collections.Generic;
 using J = Newtonsoft.Json.JsonPropertyAttribute;
@@ -33,23 +33,23 @@ namespace SeaBotCore.Data.Definitions
 
         public class Item
         {
-            [J("def_id")] public long DefId { get; set; }
-            [J("version_id")] public long VersionId { get; set; }
+            [J("def_id")] public int DefId { get; set; }
+            [J("version_id")] public int VersionId { get; set; }
             [J("name")] public string Name { get; set; }
             [J("name_loc")] public string NameLoc { get; set; }
-            [J("speedupable")] public long Speedupable { get; set; }
-            [J("sailors")] public long Sailors { get; set; }
-            [J("slots")] public long Slots { get; set; }
-            [J("travel_time")] public long TravelTime { get; set; }
-            [J("event_id")] public long EventId { get; set; }
-            [J("outpost_id")] public long OutpostId { get; set; }
-            [J("contractor_id")] public long ContractorId { get; set; }
-            [J("map_x")] public long MapX { get; set; }
-            [J("map_y")] public long MapY { get; set; }
+            [J("speedupable")] public int Speedupable { get; set; }
+            [J("sailors")] public int Sailors { get; set; }
+            [J("slots")] public int Slots { get; set; }
+            [J("travel_time")] public int TravelTime { get; set; }
+            [J("event_id")] public int EventId { get; set; }
+            [J("outpost_id")] public int OutpostId { get; set; }
+            [J("contractor_id")] public int ContractorId { get; set; }
+            [J("map_x")] public int MapX { get; set; }
+            [J("map_y")] public int MapY { get; set; }
             [J("atlas")] public string Atlas { get; set; }
             [J("texture")] public string Texture { get; set; }
-            [J("iso_width")] public long IsoWidth { get; set; }
-            [J("iso_height")] public long IsoHeight { get; set; }
+            [J("iso_width")] public int IsoWidth { get; set; }
+            [J("iso_height")] public int IsoHeight { get; set; }
             [J("dealer_texture")] public string DealerTexture { get; set; }
             [J("points")] public string Points { get; set; }
             [J("controls")] public string Controls { get; set; }
@@ -63,13 +63,13 @@ namespace SeaBotCore.Data.Definitions
 
         public class Trade
         {
-            [J("id")] public long Id { get; set; }
+            [J("id")] public int Id { get; set; }
             [J("input_type")] public string InputType { get; set; }
-            [J("input_id")] public long InputId { get; set; }
-            [J("input_amount")] public long InputAmount { get; set; }
+            [J("input_id")] public int InputId { get; set; }
+            [J("input_amount")] public int InputAmount { get; set; }
             [J("output_type")] public string OutputType { get; set; }
-            [J("output_id")] public long OutputId { get; set; }
-            [J("output_amount")] public long OutputAmount { get; set; }
+            [J("output_id")] public int OutputId { get; set; }
+            [J("output_amount")] public int OutputAmount { get; set; }
         }
     }
 }
