+++ SeaBot/Data/Definitions/OutpostDefinitions.cs
@@ -12,7 +12,7 @@
 // GNU General Public License for more details.
 // 
 // You should have received a copy of the GNU General Public License
-// along with this program.  If not, see <http://www.gnu.org/licenses/>.
+// aint with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 using System;
 using System.Collections.Generic;
@@ -39,32 +39,32 @@ namespace SeaBotCore.Data.Definitions
 
         public class Item
         {
-            [J("def_id")] public long DefId { get; set; }
-            [J("version_id")] public long VersionId { get; set; }
+            [J("def_id")] public int DefId { get; set; }
+            [J("version_id")] public int VersionId { get; set; }
             [J("name")] public string Name { get; set; }
             [J("name_loc")] public string NameLoc { get; set; }
             [J("title_loc")] public string TitleLoc { get; set; }
             [J("flavour_loc")] public string FlavourLoc { get; set; }
             [J("type")] public string Type { get; set; }
-            [J("speedupable")] public long Speedupable { get; set; }
-            [J("sailors")] public long Sailors { get; set; }
-            [J("slots")] public long Slots { get; set; }
-            [J("req_level")] public long ReqLevel { get; set; }
-            [J("travel_time")] public long TravelTime { get; set; }
-            [J("unlock_time")] public long UnlockTime { get; set; }
+            [J("speedupable")] public int Speedupable { get; set; }
+            [J("sailors")] public int Sailors { get; set; }
+            [J("slots")] public int Slots { get; set; }
+            [J("req_level")] public int ReqLevel { get; set; }
+            [J("travel_time")] public int TravelTime { get; set; }
+            [J("unlock_time")] public int UnlockTime { get; set; }
             [J("unlocked_fog")] public string UnlockedFog { get; set; }
-            [J("xp")] public long Xp { get; set; }
-            [J("xp_pct")] public long XpPct { get; set; }
-            [J("crew")] public long Crew { get; set; }
-            [J("difficulty")] public long Difficulty { get; set; }
-            [J("round")] public long Round { get; set; }
-            [J("event_id")] public long EventId { get; set; }
-            [J("map_x")] public long MapX { get; set; }
-            [J("map_y")] public long MapY { get; set; }
+            [J("xp")] public int Xp { get; set; }
+            [J("xp_pct")] public int XpPct { get; set; }
+            [J("crew")] public int Crew { get; set; }
+            [J("difficulty")] public int Difficulty { get; set; }
+            [J("round")] public int Round { get; set; }
+            [J("event_id")] public int EventId { get; set; }
+            [J("map_x")] public int MapX { get; set; }
+            [J("map_y")] public int MapY { get; set; }
             [J("atlas")] public string Atlas { get; set; }
             [J("texture")] public string Texture { get; set; }
-            [J("iso_width")] public long IsoWidth { get; set; }
-            [J("iso_height")] public long IsoHeight { get; set; }
+            [J("iso_width")] public int IsoWidth { get; set; }
+            [J("iso_height")] public int IsoHeight { get; set; }
             [J("points")] public string Points { get; set; }
             [J("controls")] public string Controls { get; set; }
