+++ SeaBot/Data/Definitions/WreckDefenitions.cs
@@ -12,7 +12,7 @@
 // GNU General Public License for more details.
 // 
 // You should have received a copy of the GNU General Public License
-// along with this program.  If not, see <http://www.gnu.org/licenses/>.
+// aint with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 using System;
 using System.Collections.Generic;
@@ -39,23 +39,23 @@ namespace SeaBotCore.Data.Definitions
 
         public class Item
         {
-            [J("def_id")] public long DefId { get; set; }
+            [J("def_id")] public int DefId { get; set; }
             [J("name")] public string Name { get; set; }
             [J("name_loc")] public string NameLoc { get; set; }
-            [J("speedupable")] public long Speedupable { get; set; }
-            [J("slots")] public long Slots { get; set; }
-            [J("enabled")] public long Enabled { get; set; }
-            [J("req_level")] public long ReqLevel { get; set; }
-            [J("travel_time")] public long TravelTime { get; set; }
-            [J("event_id")] public long EventId { get; set; }
-            [J("cooldown")] public long Cooldown { get; set; }
-            [J("xp_pct_min")] public long XpPctMin { get; set; }
-            [J("xp_pct_max")] public long XpPctMax { get; set; }
+            [J("speedupable")] public int Speedupable { get; set; }
+            [J("slots")] public int Slots { get; set; }
+            [J("enabled")] public int Enabled { get; set; }
+            [J("req_level")] public int ReqLevel { get; set; }
+            [J("travel_time")] public int TravelTime { get; set; }
+            [J("event_id")] public int EventId { get; set; }
+            [J("cooldown")] public int Cooldown { get; set; }
+            [J("xp_pct_min")] public int XpPctMin { get; set; }
+            [J("xp_pct_max")] public int XpPctMax { get; set; }
             [J("chest")] public string Chest { get; set; }
             [J("atlas")] public string Atlas { get; set; }
             [J("texture")] public string Texture { get; set; }
-            [J("iso_width")] public long IsoWidth { get; set; }
-            [J("iso_height")] public long IsoHeight { get; set; }
+            [J("iso_width")] public int IsoWidth { get; set; }
+            [J("iso_height")] public int IsoHeight { get; set; }
             [J("sets")] public Sets Sets { get; set; }
 
             [J("spots", NullValueHandling = N.Ignore)]
@@ -69,9 +69,9 @@ namespace SeaBotCore.Data.Definitions
 
         public class Set
         {
-            [J("id")] public long Id { get; set; }
-            [J("reward_count")] public long RewardCount { get; set; }
-            [J("chance")] public long Chance { get; set; }
+            [J("id")] public int Id { get; set; }
+            [J("reward_count")] public int RewardCount { get; set; }
+            [J("chance")] public int Chance { get; set; }
             [J("rewards")] public Rewards Rewards { get; set; }
         }
 
@@ -82,14 +82,14 @@ namespace SeaBotCore.Data.Definitions
 
         public class Reward
         {
-            [J("id")] public long Id { get; set; }
+            [J("id")] public int Id { get; set; }
             [J("type")] public string Type { get; set; }
-            [J("amount")] public long Amount { get; set; }
+            [J("amount")] public int Amount { get; set; }
             [J("bonus")] public double Bonus { get; set; }
             [J("koef_min")] public double KoefMin { get; set; }
             [J("koef_max")] public double KoefMax { get; set; }
-            [J("chance")] public long Chance { get; set; }
-            [J("round")] public long Round { get; set; }
+            [J("chance")] public int Chance { get; set; }
+            [J("round")] public int Round { get; set; }
         }
 
         public class Spots
@@ -99,11 +99,11 @@ namespace SeaBotCore.Data.Definitions
 
         public class Spot
         {
-            [J("id")] public long Id { get; set; }
+            [J("id")] public int Id { get; set; }
             [J("name")] public string Name { get; set; }
-            [J("map_x")] public long MapX { get; set; }
-            [J("map_y")] public long MapY { get; set; }
-            [J("outpost_id")] public long OutpostId { get; set; }
+            [J("map_x")] public int MapX { get; set; }
+            [J("map_y")] public int MapY { get; set; }
+            [J("outpost_id")] public int OutpostId { get; set; }
             [J("points")] public string Points { get; set; }
             [J("controls")] public string Controls { get; set; }
         }
