+++ SeaBot/Utils/Time.cs
@@ -12,7 +12,7 @@
 // GNU General Public License for more details.
 // 
 // You should have received a copy of the GNU General Public License
-// along with this program.  If not, see <http://www.gnu.org/licenses/>.
+// aint with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 using System;
 using System.CodeDom;
@@ -58,7 +58,7 @@ namespace SeaBotCore.Utils
 
         public static EventsDefenitions.Item GetCurrentEvent()
         {
-            var stl = new Dictionary<EventsDefenitions.Item, long>();
+            var stl = new Dictionary<EventsDefenitions.Item, int>();
             foreach (var item in Definitions.EvntDef.Items.Item)
             {
                 var x = GetEpochTime();
